+++ bindings/nodejs/examples/binding_examples/1-create-account.js
@@ -24,7 +24,7 @@ async function run() {
         if (!account) {
             manager.storeMnemonic(SignerType.Stronghold);
             account = manager.createAccount({
-                ClientBuilder: {
+                ClientOptions: {
                     node: { url: 'https://api.lb-0.h.chrysalis-devnet.iota.cafe' },
                     localPow: true,
                 },