+++ bindings/python/native/src/classes/account_manager.rs
@@ -231,7 +231,7 @@ impl AccountManager {
     }
 
     /// Adds a new account.
-    fn create_account(&self, client_options: ClientBuilder) -> Result<AccountInitialiser> {
+    fn create_account(&self, client_options: ClientOptions) -> Result<AccountInitialiser> {
         Ok(AccountInitialiser {
             account_initialiser: Some(self.account_manager.create_account(client_options.into())?),
             addresses: Default::default(),