+++ examples/wallet.rs
@@ -13,7 +13,7 @@ use iota_client::bee_message::{
 use iota_wallet::{
     account::{RemainderValueStrategy, TransferOptions},
     account_manager::AccountManager,
-    client::ClientBuilder,
+    client::ClientOptions,
     logger::{init_logger, LevelFilter},
     signing::mnemonic::MnemonicSigner,
     Result,
@@ -25,7 +25,7 @@ async fn main() -> Result<()> {
     // Generates a wallet.log file with logs for debugging
     // init_logger("wallet.log", LevelFilter::Debug)?;
 
-    let client_options = ClientBuilder::new()
+    let client_options = ClientOptions::new()
         .with_node("http://localhost:14265")?
         .with_node_sync_disabled();
 
@@ -96,7 +96,7 @@ async fn main() -> Result<()> {
     println!("Balance: {:?}", balance);
 
     // // switch to mainnet
-    // let client_options = ClientBuilder::new()
+    // let client_options = ClientOptions::new()
     //     .with_node("https://chrysalis-nodes.iota.org/")?
     //     .with_node("https://chrysalis-nodes.iota.cafe/")?
     //     .with_node_sync_disabled()
@@ -108,7 +108,7 @@ async fn main() -> Result<()> {
     // println!("Balance: {:?}", account.balance().await?);
 
     // // switch back to testnet
-    // let client_options = ClientBuilder::new()
+    // let client_options = ClientOptions::new()
     //     .with_node("https://api.lb-0.h.chrysalis-devnet.iota.cafe")?
     //     .with_node("https://api.thin-hornet-0.h.chrysalis-devnet.iota.cafe")?
     //     .with_node("https://api.thin-hornet-1.h.chrysalis-devnet.iota.cafe")?